add alternate instructions for wsl2 users. (#184)
not sure if this is the path we would like to take, but raising the question in code if helpful.  many 301 students experience a significant time-dump (for students and TAs actually) when trying to install mongoDB locally.  The answer always seem to be the same "oh, that thing we told you to do, you don't actually need to do it".  Instructions  for the "what-if" or "eventuality of" failure would preserve everyone's time and most importantly save student bandwidth for being 301 ready!  Can we just call this what it is and reap the quality of life benefits?

Alternatively, can the MongoDB install just go away?  because we allow WSL usage, using Ajax is just the way we teach them anyway!  why have half the class install mongoDB which is never used locally anyway!